fix(dashboard): 인기 가게 탑10 저장 로직 수정 (#65)
